requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
flask-jwt-extended==4.6.0
flask-cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
geopy==2.4.1
eventlet==0.35.1
